+++ Thunderstore/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2026-05-17
+
+### Changed
+
+- Dynamic Scrap Value scales from each moons base scrap values instead of current quota
+- Dynamic Scrap Amount scales from moon baseline only and no longer adds to the Dynamic Scrap Value
+
 ## [1.3.0] - 2026-05-17
 
 ### Added